# Conflicts:
#	bogenliga/bogenliga-application/src/main/java/de/bogenliga/application/services/v1/trigger/service/TriggerService.java

+++ bogenliga/bogenliga-business-logic/src/main/java/de/bogenliga/application/business/altsystem/liga/dataobject/AltsystemLigaDO.java
@@ -9,7 +9,7 @@ import de.bogenliga.application.common.altsystem.AltsystemDO;
  */
 public class AltsystemLigaDO extends AltsystemDO {
     private String name;
-    private int idNextLiga;
+    private long idNextLiga;
 
     public void setName(String name){
         this.name = name;
@@ -19,7 +19,7 @@ public class AltsystemLigaDO extends AltsystemDO {
         return this.name;
     }
 
-    public void setIdNextLiga(int idNextLiga) { this.idNextLiga = idNextLiga; }
+    public void setIdNextLiga(long idNextLiga) { this.idNextLiga = idNextLiga; }
 
-    public int getIdNextLiga(){ return this.idNextLiga; }
+    public long getIdNextLiga(){ return this.idNextLiga; }
 }
